# grassroots-backend/openAPI.json
{
  "components": {
    "schemas": {
      "ContactSearchRequestDTO": {
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "id": {
            "minimum": 1,
            "type": "number"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "CreateBulkContactRequestDTO": {
        "properties": {
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/CreateContactRequestDTO"
            },
            "type": "array"
          }
        },
        "required": ["contacts"],
        "type": "object"
      },
      "CreateContactRequestDTO": {
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "organizationId": {
            "minimum": -1,
            "type": "number"
          },
          "phoneNumber": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "phoneNumber",
          "organizationId"
        ],
        "type": "object"
      },
      "CreateOrganizationNoParentRequestDTO": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": ["name"],
        "type": "object"
      },
      "CreateOrganizationRequestDTO": {
        "properties": {
          "name": {
            "type": "string"
          },
          "parentID": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": ["name", "parentID"],
        "type": "object"
      },
      "LoginStateDTO": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDTO"
          }
        },
        "type": "object"
      },
      "OrganizationReferenceDTO": {
        "properties": {
          "id": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": ["id"],
        "type": "object"
      },
      "PaginatedContactSearchRequestDTO": {
        "properties": {
          "contact": {
            "$ref": "#/components/schemas/ContactSearchRequestDTO"
          },
          "paginated": {
            "$ref": "#/components/schemas/PaginatedRequestDTO"
          }
        },
        "required": ["contact", "paginated"],
        "type": "object"
      },
      "PaginatedRequestDTO": {
        "properties": {
          "rowsToSkip": {
            "minimum": 0,
            "type": "number"
          },
          "rowsToTake": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": ["rowsToSkip", "rowsToTake"],
        "type": "object"
      },
      "RoleDTO": {
        "properties": {
          "id": {
            "minimum": 1,
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": ["VIEW_CONTACTS", "MANAGE_CONTACTS", "MANAGE_USERS"],
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": ["permissions", "id", "name"],
        "type": "object"
      },
      "UserDTO": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "emails": {
            "items": {
              "format": "email",
              "type": "string"
            },
            "type": "array"
          },
          "firstName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "userRoles": {
            "items": {
              "$ref": "#/components/schemas/UserRoleDTO"
            },
            "type": "array"
          }
        },
        "required": ["id"],
        "type": "object"
      },
      "UserRoleDTO": {
        "properties": {
          "id": {
            "minimum": 1,
            "type": "number"
          },
          "inherited": {
            "type": "boolean"
          },
          "organizationId": {
            "minimum": 1,
            "type": "number"
          },
          "role": {
            "$ref": "#/components/schemas/RoleDTO"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["role", "organizationId", "inherited"],
        "type": "object"
      },
      "ValidationErrorOutDTO": {
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "statusCode": {
            "type": "number"
          }
        },
        "required": ["statusCode", "message", "error"],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "The Grassroots API description",
    "title": "Grassroots",
    "version": "0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["App"]
      }
    },
    "/auth/active-org": {
      "get": {
        "operationId": "AuthController_getActiveOrg",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/example_route_using_user": {
      "get": {
        "operationId": "AuthController_example",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginStateDTO"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/google/callback": {
      "get": {
        "operationId": "AuthController_googleAuthRedirect",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/is_authenticated": {
      "get": {
        "operationId": "AuthController_isUserLoggedIn",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/login": {
      "get": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "in": "query",
            "name": "redirect_path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/set-active-org": {
      "post": {
        "operationId": "AuthController_setActiveOrg",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationReferenceDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Auth"]
      }
    },
    "/contacts": {
      "get": {
        "operationId": "ContactsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Contacts"]
      },
      "post": {
        "operationId": "ContactsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Contacts"]
      }
    },
    "/contacts/bulk-create": {
      "post": {
        "operationId": "ContactsController_bulkCreate",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBulkContactRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Contacts"]
      }
    },
    "/contacts/search": {
      "post": {
        "operationId": "ContactsController_search",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginatedContactSearchRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Contacts"]
      }
    },
    "/contacts/{id}": {
      "get": {
        "operationId": "ContactsController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Contacts"]
      }
    },
    "/organizations": {
      "get": {
        "operationId": "OrganizationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Organizations"]
      },
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Organizations"]
      }
    },
    "/organizations/ancestors-of/{id}": {
      "get": {
        "operationId": "OrganizationsController_getAncestors",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Organizations"]
      }
    },
    "/organizations/create-root": {
      "post": {
        "operationId": "OrganizationsController_createRoot",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationNoParentRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Organizations"]
      }
    },
    "/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Organizations"]
      }
    },
    "/roles": {
      "get": {
        "operationId": "RolesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Roles"]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Users"]
      }
    },
    "/users/find-or-create": {
      "post": {
        "operationId": "UsersController_findOrCreate",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Users"]
      }
    },
    "/users/user-permissions-for-org": {
      "get": {
        "operationId": "UsersController_getUserPermissionsForOrg",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "organizationId",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOutDTO"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "tags": ["Users"]
      }
    }
  },
  "servers": [],
  "tags": []
}